I've come to the realization that collision detection is relative to the application; it's very easy to get traped with edge case after edge case, and ultimately collsion detection/resolution is best designed for the particular simulation. In short, screw it: Tiles.
Tiles keep track of the sprites around them (N, E, S, W); if the  ideal delta is pushing the player in to another tile, then instead it resolves to the other delta. Also now Tiles must be uniform (e.g. 16x16) and bound on the same grid. It's not the grand result I'd envisioned, but the illusion is close enough.

Need to write some tests and clean up, but otherwise collision detection is looking good.