Implemented the simple MoveLeft behavior, and it's looking good. Now I just need to add the hitbox check for bullets vs vehicles and I'll have something resembling a game!
Also, I realized an issue in the way I was handling offsets; that is, the offset step is multiplied 16. Looks like I'll have to add another x/y offset property that will shift the hitboxes away from always being 0,0.

Also also, there may be a bug in the MoveSprite call, but I'll deal with that when I see it again.