Badabang, we have Illusions, the newest member to the Sprite family! Illusions can be spawned similar to bullets, and may only have hitboxes if they're subject to gravity. Right now, when an enemy tank explodes, it spawns an explosion illusion wherever it is (falling to only be caught by a tile), and when it is finished it cleans itself up.
One notable change that was also done with this, is moving the UpdateActionSequence call out of all calling updaters to the main composite, ensuring it gets called last. this is because sprites can be destroyed, and thus removed from the grid by the update action sequence call. Now, by placing it in the successful condition of the composite, it is ensured to A. always be called for every sprite exactly once, and B. it can safely remove a sprite from a grid without having to communicate that up to the updater.